+++ src/main/java/es/in2/issuer/backend/shared/domain/util/GlobalErrorTypes.java
@@ -50,7 +50,8 @@ public enum GlobalErrorTypes {
     PARSE_ERROR_EXCEPTION("parse_error_exception"),
     TOKEN_FETCH_ERROR("token_fetch_error"),
     WELL_KNOWN_INFO_FETCH_ERROR("well_known_info_fetch_error"),
-    INVALID_CREDENTIAL_STATUS_TRANSITION("invalid_credential_status_transition");
+    INVALID_CREDENTIAL_STATUS_TRANSITION("invalid_credential_status_transition"),
+    UNKNOWN_CREDENTIAL_CONFIGURATION("unknown_credential_configuration");
 
     private final String code;